Encapsulated greeting and exiting into processor
