+++ src/main/java/Processor.java
@@ -1,6 +1,18 @@
 import java.util.Objects;
 
 public class Processor {
+    private static final String LOGO_PATH = "./src/main/Snorlax.txt";
+    private static final String GREETING_PATH = "./src/main/Greeting.txt";
+    private static final String EXIT_PATH = "./src/main/Exit.txt";
+
+    public static void greet() {
+        String logo = Utils.getFile(LOGO_PATH);
+        String greeting = Utils.getFile(GREETING_PATH);
+
+        System.out.println(logo);
+        Utils.printLine();
+        System.out.println(greeting);
+    }
     public static void process (String input, TaskList tasks) {
         input = input.trim();
         String command = Utils.firstWord(input);
@@ -27,5 +39,10 @@ public class Processor {
             default:
                 Utils.encaseLines("Invalid command");
         }
+    }
+
+    public static void exit() {
+        String exit = Utils.getFile(EXIT_PATH);
+        Utils.encaseLines(exit);
     }
 }
