+++ text-ui-test/input.txt
@@ -1,3 +1,4 @@
-read book
+todo read book
 list
 mark 1
+bye
